moved proc-macro code into a derive project and restructured workspace; repo is building
