+++ rust/Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "gen-cpp-doc",
     "md2man",
     "psibase_macros",
+    # "psibase-macros-derive",
     "psibase",
     "test_contract",
     "test_contract_2",
@@ -16,7 +17,7 @@ members = [
     "test_service_elections",
     "test_service_sites",
     "test_fracpack",
-    "test_subjective"
+    "test_subjective",
 ]
 
 default-members = [
@@ -27,7 +28,7 @@ default-members = [
     "md2man",
     "psibase_macros",
     "psibase",
-    "test_fracpack"
+    "test_fracpack",
 ]
 
 [workspace.package]
