+++ rust/psibase_macros/Cargo.toml
@@ -8,13 +8,5 @@ edition = "2021"
 description = "Macros to support psibase development"
 license = "MIT"
 
-[lib]
-proc-macro = true
-
 [dependencies]
-darling = "0.14"
-proc-macro-error = "1.0.4"
-proc-macro2 = "1.0.36"
-psibase_names = { version = "0.13.0", path = "../psibase_names" }
-quote = "1.0.15"
-syn = {version = "1.0.86", features = ["full", "visit-mut"] }
+psibase-macros-derive = { path = "psibase-macros-derive" }
