+++ rust/psibase_macros/psibase-macros-derive/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "psibase-macros-derive"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+darling = "0.14"
+proc-macro-error = "1.0.4"
+proc-macro2 = "1.0.36"
+psibase_names = { version = "0.12.0", path = "../../psibase_names" }
+quote = "1.0.15"
+syn = { version = "1.0.86", features = ["full", "visit-mut"] }
