+++ rust/psibase_macros/psibase-macros-derive/src/schema_macro.rs
